requirements_biomedclip.txt
# BiomedCLIP 乳腺良/恶性分类
torch>=1.12.0
torchvision>=0.13.0
open_clip_torch>=2.23.0
transformers>=4.35.0
Pillow>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0
tqdm>=4.60.0
